review fixes: assert 8037 gas invariants, restore getLogs depth test, fix stale docs
- Debug.Assert the EIP-8037 boundary invariants (state >= 0, remaining +
  reservoir <= gasLimit) at the halt/fail/collision read points and in
  RefundStateGas, replacing the saturating defense removed with the rework
- restore Eth_get_logs_enforces_max_block_depth (accidentally dropped in the
  master RPC reconciliation; the MaxBlockDepth guard is live)
- de-circularize the 7702 SpentGas expectation (constants instead of the
  production intrinsic calculator) and add an exact floor==standard tie case
  for EIP-7981
- refresh stale 4500-era totals/comments (TX_BASE is 12000), the 8038
  placeholder remark, and the 8037 inclusion-rule assertion message
- drop unused Eip8038IntrinsicRecipientGas params, redundant (ulong) casts on
  ulong constants, and the lazy BlockValidator ecdsa init

+++ src/Nethermind/Nethermind.Evm/GasPolicy/IGasPolicy.cs
@@ -189,8 +189,8 @@ public interface IGasPolicy<TSelf> where TSelf : struct, IGasPolicy<TSelf>
 
         (int addressesCount, int storageKeysCount) = accessList.Count;
         // EIP-8038 realigns access-list entry costs with the cold-access costs they pre-warm.
-        ulong addressCost = (ulong)(spec.IsEip8038Enabled ? Eip8038Constants.AccessListAddressCost : GasCostOf.AccessAccountListEntry);
-        ulong storageKeyCost = (ulong)(spec.IsEip8038Enabled ? Eip8038Constants.AccessListStorageKeyCost : GasCostOf.AccessStorageListEntry);
+        ulong addressCost = spec.IsEip8038Enabled ? Eip8038Constants.AccessListAddressCost : GasCostOf.AccessAccountListEntry;
+        ulong storageKeyCost = spec.IsEip8038Enabled ? Eip8038Constants.AccessListStorageKeyCost : GasCostOf.AccessStorageListEntry;
         return (ulong)addressesCount * addressCost
             + (ulong)storageKeysCount * storageKeyCost
             + spec.GasCosts.TotalCostFloorPerToken * floorTokensInAccessList;
@@ -215,7 +215,7 @@ public interface IGasPolicy<TSelf> where TSelf : struct, IGasPolicy<TSelf>
 
         ulong authCount = (ulong)authList.Length;
         // EIP-8038 reprices the per-authorization regular cost (ACCOUNT_WRITE + auth-base).
-        ulong perAuthRegular = (ulong)(spec.IsEip8038Enabled ? Eip8038Constants.PerAuthBaseRegular : GasCostOf.PerAuthBaseRegular);
+        ulong perAuthRegular = spec.IsEip8038Enabled ? Eip8038Constants.PerAuthBaseRegular : GasCostOf.PerAuthBaseRegular;
         return spec.IsEip8037Enabled
             ? (
                 authCount * perAuthRegular,
